Add new Delegated Authentication Protocol
Authentication is delegated by the Strategy to an external provider
but handled locally (e.g. makes request internally to external provider)
unlike OAuth or similar which redirects to the external provider.
On success the authenticated user is connected to a local user (created
if necessary) in a similar manner to OAuth.

Add new POST route to support delegated authentication, which requires
the ability to post a login form in a manner similar to the local
protocol.